move BottomSheet component to folder
