+++ src/modules/stackables/components/bottomSheet/BottomSheet.js
@@ -3,9 +3,9 @@
  */
 import { html, nothing } from 'lit-html';
 import css from './bottomSheet.css';
-import { MvuElement } from '../../MvuElement';
-import { closeBottomSheet } from '../../../store/bottomSheet/bottomSheet.action';
-import closeIcon from './assets/x-square.svg';
+import { MvuElement } from '../../../MvuElement';
+import { closeBottomSheet } from '../../../../store/bottomSheet/bottomSheet.action';
+import closeIcon from '../assets/x-square.svg';
 
 const Update = 'update';
 const Update_Main_Menu = 'update_main_menu';
